% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addQCfilter.R
\name{addQCfilter}
\alias{addQCfilter}
\title{Add QC filters}
\usage{
addQCfilter(
  obj,
  filterName = "input",
  by_threshold = FALSE,
  cutoffs = defaultCutoffs,
  by_outlier = FALSE,
  nmads = 4
)
}
\arguments{
\item{obj}{Seurat object}

\item{filterName}{name of new metadata column with logical filter}

\item{by_threshold}{logical to control whether to filter on thresholds defined in 'cutoffs'}

\item{cutoffs}{named list of cutoffs}

\item{by_outlier}{logical to control whether to filter outliers}

\item{nmads}{median absolute deviation used to define outliers}
}
\value{
Seurat object with new metadata column
}
\description{
Adds logical metadata column for a given set of arbitrary cutoffs and/or outliers
}
