% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addQCmetrics.R
\name{addQCmetrics}
\alias{addQCmetrics}
\title{Add QC metrics}
\usage{
addQCmetrics(obj, mito.pattern = "^MT", ribo.pattern = "^RP[SL]")
}
\arguments{
\item{obj}{Seurat object}

\item{mito.pattern}{pattern defining mitochondiral genes}

\item{ribo.pattern}{pattern defining ribosomal genes}
}
\value{
Seurat object with modified metadata
}
\description{
Calculates qc metrics and reformats Seurat object metadata to be more readable
}
