% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCC.R
\name{checkCC}
\alias{checkCC}
\title{Perform cell cycle check analysis}
\usage{
checkCC(
  obj,
  what2return = c("plot_list", "seurat_obj", "both"),
  cc.genes,
  features2plot = c("Phase", "mitoRatio", "riboRatio")
)
}
\arguments{
\item{obj}{Seurat object}

\item{what2return}{c("plot_list", "seurat_obj", "both")}

\item{cc.genes}{Vector of gene names for cell cycle genes. If you are starting with Ensembl IDs use gencoreSC::EnsDb2GeneName() first.}

\item{features2plot}{Vector of categorical features to plot PCA for e.g. c("Phase", "mitoRatio", "riboRatio"))}
}
\value{
Seurat object with Phase metadata column, a plot list with PCAs of features specified by features2plot, or both in a list object.
}
\description{
Calculates phase of cell cycle given a list of cell cycle genes
}
\note{
To do:
\itemize{
\item Fix namesopace warnings and notes from using methods defined by AnnotationHub
}
}
