% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCC.R
\name{checkPCA}
\alias{checkPCA}
\title{Plot PCAs for a list of categorical features}
\usage{
checkPCA(obj, features = c("Phase", "mitoRatio", "riboRatio"))
}
\arguments{
\item{obj}{Seurat object}

\item{features}{Vector of features to to group.by and split.by in PCA plots}
}
\value{
List of plot objects
}
\description{
Take a vector of features and plot PCA plots for each, returning a list.
}
