% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSingleR.R
\name{clID2clNames}
\alias{clID2clNames}
\title{Convert cell ontology IDs to names}
\usage{
clID2clNames(
  metadata,
  cl,
  old.prefix = "ImmGen.ont",
  new.prefix = "ImmGen.ont.name"
)
}
\arguments{
\item{metadata}{Data.frame with labels, pruned.labels and delta.next columns. Typically either:
\itemize{
\item Seurat object metadata (after running SingleR), or
\item SingleR output object
}}

\item{cl}{Cell ontology database
\itemize{
\item Can be btained via: \code{cl <- ontoProc::getOnto('cellOnto')}
}}

\item{old.prefix}{The prefix used for the original runSingleR call (e.g. "ImmGen.ont")}

\item{new.prefix}{The new prefix to use for the human readable column (e.g. "ImmGen.ont.name")}
}
\value{
List with obj.seurat and SingleR output. Note: should the SingleR just be in misc. slot?
}
\description{
Convert bare cell ontology IDs to human-readable names
}
\examples{
\dontrun{
## Run on bare ontology IDs, then convert to human readable cell type labels
# Get cell ontology names and other info
cl <- ontoProc::getOnto('cellOnto')

# ImmGen.ont
list[obj.seurat, SingleR.ImmGen.ont] <- obj.seurat \%>\%
  Run.SingleR(obj.seurat=.,
              ref=ref.ImmGen,
              labels=ref.ImmGen$label.ont,
              de.method="classic",
              meta.prefix = "ImmGen.ont")

# These labels aren't very informative:
SingleR::plotScoreHeatmap(SingleR.ImmGen.ont, show.pruned = TRUE, fontsize=4)

# Convert bare cell ontology IDs to informative names in seurat object
obj.seurat@meta.data <- clID2clNames(obj.seurat@meta.data, cl,
                                     old.prefix = "ImmGen.ont", new.prefix = "ImmGen.ont.name")

# Convert bare cell ontology IDs to informative names in SingleR output and visualize
SingleR.ImmGen.ont.names <- clID2clNames(SingleR.ImmGen.ont, cl,
                                         old.prefix = "ImmGen.ont", new.prefix = "ImmGen.ont.name")
SingleR::plotScoreHeatmap(SingleR.ImmGen.ont.names, show.pruned = TRUE, fontsize=4)
}

}
