% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSingleR.R
\name{convertClID2Name}
\alias{convertClID2Name}
\title{Convert cell ontology IDs to names in SingleR output}
\usage{
convertClID2Name(
  metadata,
  cl,
  old.prefix = "ImmGen.ont",
  new.prefix = "ImmGen.ont.name"
)
}
\arguments{
\item{metadata}{data.frame with labels, pruned.labels and delta.next columns. Typically either:
\itemize{
\item Seurat object metadata (after running SingleR), or
\item SingleR output object
}}

\item{cl}{Cell ontology database
\itemize{
\item Can be btained via: \code{cl <- ontoProc::getOnto('cellOnto')}
}}

\item{old.prefix}{The prefix used for the original runSingleR call (e.g. "ImmGen.ont")}

\item{new.prefix}{The new prefix to use for the human readable column (e.g. "ImmGen.ont.name")}
}
\value{
data.frame with three new columns: \verb{\{new.prefix\}.labels}, \verb{\{new.prefix\}.pruned.labels}, \verb{\{new.prefix\}.delta.next}.
}
\description{
Convert bare cell ontology IDs used in SingleR to human-readable names
}
\examples{
\dontrun{
## Run on bare ontology IDs, then convert to human readable cell type labels
# Get cell ontology names and other info
cl <- ontoProc::getOnto('cellOnto')

# ImmGen.ont
c(obj.seurat, SingleR.ImmGen.ont) \%<-\% obj.seurat \%>\%
  Run.SingleR(obj.seurat=.,
              ref=ref.ImmGen,
              labels=ref.ImmGen$label.ont,
              de.method="classic",
              meta.prefix = "ImmGen.ont")

# These labels aren't very informative:
SingleR::plotScoreHeatmap(SingleR.ImmGen.ont, show.pruned = TRUE, fontsize=4)

# Convert bare cell ontology IDs to informative names in seurat object
obj.seurat@meta.data <- clID2clNames(metadata = obj.seurat@meta.data, cl = cl,
                                     old.prefix = "ImmGen.ont", new.prefix = "ImmGen.ont.name")

# Convert bare cell ontology IDs to informative names in SingleR output and visualize
SingleR.ImmGen.ont.names <- clID2clNames(metadata = SingleR.ImmGen.ont, cl = cl,
                                         old.prefix = "ImmGen.ont", new.prefix = "ImmGen.ont.name")

SingleR::plotScoreHeatmap(SingleR.ImmGen.ont.names, show.pruned = TRUE, fontsize=4)
}

}
