% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demuxAntibodyData.R
\name{demuxAntibodyData}
\alias{demuxAntibodyData}
\title{Wrapper for calling HTODemux on antibody capture data in a Seurat object}
\usage{
demuxAntibodyData(
  obj,
  labels,
  assay = "Antibody.Capture",
  normalization.method = "CLR",
  ...
)
}
\arguments{
\item{obj}{A Seurat object with antibody capture / hashtag oligo capture data}

\item{labels}{A named character vector, where values are the desired labels
and the names are the HTO features as output by Cellranger}

\item{assay}{Name of assay to pull data from}

\item{normalization.method}{Method for normalization.
 \itemize{
  \item{LogNormalize: }{Feature counts for each cell are divided by the total
  counts for that cell and multiplied by the scale.factor. This is then
  natural-log transformed using log1p.}
  \item{CLR: }{Applies a centered log ratio transformation}
  \item{RC: }{Relative counts. Feature counts for each cell are divided by the total
  counts for that cell and multiplied by the scale.factor. No log-transformation is applied.
  For counts per million (CPM) set \code{scale.factor = 1e6}}
}}

\item{...}{
  Arguments passed on to \code{\link[Seurat:HTODemux]{Seurat::HTODemux}}
  \describe{
    \item{\code{positive.quantile}}{The quantile of inferred 'negative' distribution for each hashtag - over which the cell is considered 'positive'. Default is 0.99}
    \item{\code{init}}{Initial number of clusters for hashtags. Default is the # of hashtag oligo names + 1 (to account for negatives)}
    \item{\code{nstarts}}{nstarts value for k-means clustering (for kfunc = "kmeans"). 100 by default}
    \item{\code{kfunc}}{Clustering function for initial hashtag grouping. Default is "clara" for fast k-medoids clustering on large applications, also support "kmeans" for kmeans clustering}
    \item{\code{nsamples}}{Number of samples to be drawn from the dataset used for clustering, for kfunc = "clara"}
    \item{\code{seed}}{Sets the random seed. If NULL, seed is not set}
    \item{\code{verbose}}{Prints the output}
  }}
}
\value{
The Seurat object with the following demultiplexed information stored in the meta data:
\describe{
  \item{hash.maxID}{Name of hashtag with the highest signal}
  \item{hash.secondID}{Name of hashtag with the second highest signal}
  \item{hash.margin}{The difference between signals for hash.maxID and hash.secondID}
  \item{classification}{Classification result, with doublets/multiplets named by the top two highest hashtags}
  \item{classification.global}{Global classification result (singlet, doublet or negative)}
  \item{hash.ID}{Classification result where doublet IDs are collapsed}
}
}
\description{
Provide a list of sample labels and hashtag oligos as a named list,
where the name is the antibody capture feature and the value is the
desired label. The HTO data will be extracted into it's own assay
(in case there is other antibody data in the original assay), have the
feature names converted to the desired labels, and run \code{Seurat::HTODemux()}.
}
\examples{
\dontrun{
##> hash_labels
  objs_filt <- lapply(objs_filt, demuxAntibodyData, labels = hash_labels)
}
}
