% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCC.R
\name{EnsDb2GeneName}
\alias{EnsDb2GeneName}
\title{Convert cell cycle Ensembl IDs to gene names}
\usage{
EnsDb2GeneName(cell_cycle_genes, species = "Mus musculus")
}
\arguments{
\item{cell_cycle_genes}{Vector of Ensemble IDs for cell cycle genes}

\item{species}{Binomial species epithet for organism (default = "Mus musculus")}
}
\value{
Vector of cell cycle gene names
}
\description{
Convert a vector of cell cycle Ensembl IDs to gene symbols
}
