% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGeneCorrectionsSoupX.R
\name{extractGeneCorrectionsSoupX}
\alias{extractGeneCorrectionsSoupX}
\title{Extract SoupX contamination correction values for a gene}
\usage{
extractGeneCorrectionsSoupX(gene, sc)
}
\arguments{
\item{gene}{A gene symbol}

\item{sc}{SoupChannel object produced by \code{runSoupX()}}
}
\value{
A list of named numeric values
}
\description{
Pull unmodified transcript counts, soupX corrected counts, portion of
cells set to 0 expression, and soup profile counts for a single gene
from a modified SoupChannel produced by \code{runSoupX()}. This can be used
to see if a gene of interest from a DGE or similar analysis had strong
presence in the background contamination and may be influenced by SoupX
correction.
}
\examples{
\dontrun{
sc <- runSoupX(tod = tod,
  toc = toc,
  clus = clusters)
extractGeneCorrectionsSoupX('ZFPM2', sc)
}
}
