% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRefMapTopResults.R
\name{extractTopRefMapResults}
\alias{extractTopRefMapResults}
\title{Extract the top N cell-type annotations by cluster}
\usage{
extractTopRefMapResults(
  obj,
  top_n,
  label_column,
  label_score_column,
  mapping_score_column = NULL,
  clusters_column = "seurat_clusters"
)
}
\arguments{
\item{obj}{A Seurat object with cell-type annotations and scores from something
like Azimuth or SingleR}

\item{top_n}{How many annotations to report for each cluster}

\item{label_column}{Metadata column to extract cell-type labels from}

\item{label_score_column}{Metadata column to extract annotation scores from}

\item{mapping_score_column}{Optional, column to extract mapping scores from
(from Azimuth)}

\item{clusters_column}{Column to extract clusters from}
}
\value{
A dataframe of \code{top_n} rows per cluster
}
\description{
Extract the top N cell-type annotations for each cluster by number of cells.
For each annotation per cluster the median confidence score is reported.
The output can optionally include the median mapping score from Azimuth outputs.
}
\examples{
\dontrun{
extractRefMapTopResults(combined.obj,
label_column = 'predicted.cell_type.lymphTS',
label_score_column = 'predicted.cell_type.score.lymphTS',
mapping_score_column = 'mapping.score.lymphTS',
top_n = 2) \%>\%
  knitr::kable(digits = 3,
               col.names = c('Cluster',
                             'Predicted cell type',
                             'Median label confidence score',
                             'Median mapping confidence score',
                             'Proportion of cluster'),
               caption = 'Top 2 most common calls for each cluster') \%>\%
  kable_styling(full_width = F)

}
}
