% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureHeatmapByCluster.R
\name{featureHeatmapByCluster}
\alias{featureHeatmapByCluster}
\title{featureHeatmapByCluster}
\usage{
featureHeatmapByCluster(
  obj,
  assay,
  features = NULL,
  main = obj@project.name,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  n_features_to_plot = 50,
  cluster_column = "seurat_clusters",
  ...
)
}
\arguments{
\item{obj}{A Seurat object}

\item{assay}{Assay to pull feature expression data from}

\item{features}{Optional: specify which features to plot. If none provided,
defaults to plotting first \code{n_features_to_plot}}

\item{main}{the title of the plot}

\item{cluster_rows}{boolean values determining if rows should be clustered or \code{hclust} object,}

\item{cluster_cols}{boolean values determining if columns should be clustered or \code{hclust} object.}

\item{n_features_to_plot}{Number of features to plot if no features are provided.
Default 50 to avoid accidentally attempting to plot all features.}

\item{cluster_column}{Metadata column containing cluster identities. Default
is \code{seurat_clusters}}

\item{\dots}{Additional arguments passed to pheatmap::pheatmap}
}
\value{
A ggplot converted from a pheatmap
}
\description{
Create a heatmap of feature expression by cluster. Specify 'assay' to
pull either RNA or ADT data. If no features are provided, the first
'n_features_to_plot' (default 50) will be used. This is to make it easy
to quickly plot all ADT data, while also making this function usable
with RNA data, but not breaking your computer by trying to accidentally
render all RNA features.
}
\examples{
\dontrun{
  featureHeatmapByCluster(obj.myeloids, assay = 'ADT', features = c('CD27.1', 'IgD', 'CD11c'))
}
}
