% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLowGenes.R
\name{filterLowGenes}
\alias{filterLowGenes}
\title{Remove genes present in low number of cells}
\usage{
filterLowGenes(obj, min.cells, calculate_only = TRUE)
}
\arguments{
\item{obj}{Seurat object}

\item{min.cells}{Minimum number of cells a gene must appear in to be kept}

\item{calculate_only}{If \code{TRUE} only show how many genes would be removed,
but return the unfiltered object. If \code{FALSE}, return the filtered object}
}
\value{
An object of class Seurat
}
\description{
Remove rows from Seurat object for genes present in fewer than \code{min.cells}
cells. This is normally done in \code{SeuratObject::CreateSeuratObject()}, but separating
this out allows SoupX to be added to the workflow. After SoupX processing,
this function can be ran to filter the object as you would normally.
}
\examples{
\dontrun{
  ## See the impact of filtering
  objs <- lapply(objs, filterLowGenes, min.cells = 100)
  ## apply filter
  objs <- lapply(objs, filterLowGenes, min.cells = 100, calculate_only = FALSE)
}
}
