% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSoupXFiles.R
\name{findSoupXFiles}
\alias{findSoupXFiles}
\title{Find SoupX input files in Cellranger output folder}
\usage{
findSoupXFiles(base_path)
}
\arguments{
\item{base_path}{Path to cellranger outputs}
}
\value{
A list of file paths
}
\description{
Cellranger has different outputs based on what is ran, so this function
should help locate the necessary files for SoupX from the labrynth of output
folders. It has not been robustly tested with multiple Cellranger output
types, so it might not be general enough for other runs, but hopefully
there is at least some consistency in outputs.
}
\examples{
\dontrun{

## This returns all soupx files for all samples in the samplesheet
soupx_files <- lapply(samples$FileID, function(x){
  findSoupXFiles(file.path(config$rootDir, config$alignmentDir, x))
})
}
}
