% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addQCfilter.R
\name{generate_capture_QC_cutoffs}
\alias{generate_capture_QC_cutoffs}
\title{generate_capture_QC_cutoffs}
\usage{
generate_capture_QC_cutoffs(
  capture,
  nmads = 4,
  metrics = c("nUMI", "nGene", "riboRatio", "mitoRatio", "log10GenesPerUMI")
)
}
\arguments{
\item{capture}{A seurat object}

\item{nmads}{Number of median absolute deviations, default 4}

\item{metrics}{Columns of the capture metadata to calculate thresholds for}
}
\value{
a list of min and max thresholds for each of \code{metrics}
}
\description{
Generates default cutoffs for all metrics of interest based on median
absolute deviation for the distribution of metric values in the capture.
}
\examples{
\dontrun{
  cutoffs <- generate_cutoffs(objs$Apr05_HH_5A)
  ## multiple captures
  cutoffs <- lapply(objs, generate_cutoffs)
}
}
