% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopNGenesSC.R
\name{getTopNGenesSC}
\alias{getTopNGenesSC}
\title{Return top genes from Seurat::FindMarkers result}
\usage{
getTopNGenesSC(
  result,
  N = 30,
  min_logFC = log2(1.5),
  min_padj = 0.05,
  min_pct = 0.25,
  exclusion_patterns = c("^ENS", "^MT", "^RP[SL]"),
  direction = "mixed",
  lfc_col = "avg_log2FC",
  pval_col = "p_val_adj"
)
}
\arguments{
\item{result}{data.frame from \code{Seurat::FindMarkers} or similar function}

\item{N}{Number of genes to return, max}

\item{min_logFC}{Minimum absolute value for logFC}

\item{min_padj}{Minimum adjusted p value}

\item{min_pct}{Minimum percent expressed}

\item{exclusion_patterns}{Patterns to exclude. Defaults to ENS IDs, mitochondrial
genes, and ribosomal genes}

\item{direction}{one of "mixed, up, down, equal". Dictates the direction
of expression for returned genes. 'Mixed' does no filtering based on direction.
Up and down specify positive and negative log FC. Equal returns an equal
number of up and down regulated genes.}

\item{lfc_col}{Column containing log-fold change values}

\item{pval_col}{Column containing adjusted p values}
}
\value{
A vector of gene names
}
\description{
Returns top N genes from a \code{Seurat::Findmarkers} or other differential expression
result, based on some optional filtering criteria. Mainly intended for quick
returns of interesting genes to plot.
}
\examples{
\dontrun{
## get the top diff. expressed genes, then plot them

gcoreVlnPlot(myeloid_obj,
 genes = getTopNGenesSC(seurat_results$global_overall, min_logFC = 0.4, N = 6),
 grouping_var = 'stage',
 groups = c('Pre', 'Post'))
}
}
