% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIntegration.R
\name{integrate_harmony}
\alias{integrate_harmony}
\title{Run integration via Harmony}
\usage{
integrate_harmony(
  s,
  norm_method = "logNorm",
  nfeatures = 2000,
  npcs = 30,
  group.by.vars = "Sample_Name",
  verbose = FALSE,
  regress.out = NULL,
  feature.blacklist = NULL,
  scale_data = NULL,
  reduction.save = "harmony"
)
}
\arguments{
\item{s}{A Seurat object. If it is a split object, will run normalize samples separately. If merged, will normalize samples together.}

\item{norm_method}{Normalization method (either "logNorm" = log transform or "SCT" = SCTransform)}

\item{nfeatures}{Number of features to pass to Seurat::FindVariableFeatures()}

\item{npcs}{Total Number of PCs to compute and store (30 by default)}

\item{group.by.vars}{the name(s) of covariates that harmony will remove
its effect on the data.}

\item{verbose}{enable verbosity}

\item{regress.out}{Features to regress out in SCTransform or ScaleData vars.to.regress}

\item{feature.blacklist}{Vector of genes to blacklist from variable features}

\item{scale_data}{Logical; whether to run Seurat::ScaleData(). If NULL (default), will scale for log normalization but not for SCTransform.}

\item{reduction.save}{the name of the new slot that is going to be created by
harmony. By default, HARMONY.}
}
\value{
Seurat (v3) object. Harmony dimensions placed into dimensional reduction object specified by "reduction.save" (default = "harmony"). For downstream Seurat analyses, use reduction='harmony'. See \code{\link[harmony:RunHarmony]{harmony::RunHarmony}} for details.
The object also includes PCA and UMAP reductions and neighbor graphs and seurat clusters defined at res=0.6.
}
\description{
A wrapper function for integration of RNA assays via \code{\link[harmony:RunHarmony]{Harmony}}; normalizing, regressing out or blacklisting features etc. as specified. The intention is to streamline analyses where multiple integration methods or parameter combinations need to be compared. Thus, the input Seurat object need only contain raw RNA counts.
}
\examples{
\dontrun{
# log normalize each capture separately and integrate via Harmony
# s.split is a Seurat object split by "Sample_Name"
s.Harmony <- integrate_harmony(s.split, norm_method = "logNorm", group.by.vars = "Sample_Name)

# SCTransform each capture separately and integrate via Harmony
s.Harmony <- integrate_harmony(s, norm_method = "SCT", group.by.vars = "Sample_Name)

# SCTransform each capture separately, blacklist TCR genes from variable features
library(scGate)
TCR_genes <- scGate::genes.blacklist.default$Mm$TCR
s.Harmony <- integrate_harmony(s, norm_method = "SCT", group.by.vars = "Sample_Name,
                               feature.blacklist = "TCR_genes")

# SCTransform all captures together and integrate via Harmony
s <- merge(x = s.split[[1]], y = s.split[2:length(s.split)], merge.data = FALSE)
s.Harmony <- integrate_harmony(s, norm_method = "SCT", group.by.vars = "Sample_Name)
}

}
