% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIntegration.R
\name{integrate_seurat}
\alias{integrate_seurat}
\title{Run integration via Seurat}
\usage{
integrate_seurat(
  s.split,
  dim.reduct = "cca",
  nfeatures = 2000,
  npcs = 30,
  verbose = FALSE,
  norm_method = "logNorm",
  feature.blacklist = NULL,
  regress.out = NULL,
  k.anchor = 5
)
}
\arguments{
\item{s.split}{A split Seurat object containing all samples/captures to integrate}

\item{dim.reduct}{Reduction to use for Seurat integration ("cca" or "rpca"). see \code{\link[Seurat:FindIntegrationAnchors]{Seurat::FindIntegrationAnchors}} reduction parameter for details.}

\item{nfeatures}{Number of features to pass to Seurat::FindVariableFeatures()}

\item{npcs}{Total Number of PCs to compute and store (30 by default)}

\item{verbose}{Print progress bars and output}

\item{norm_method}{Normalization method (either "logNorm" = log transform or "SCT" = SCTransform)}

\item{feature.blacklist}{Vector of genes to blacklist from variable features}

\item{regress.out}{Features to regress out in SCTransform vars.to.regress}

\item{k.anchor}{How many neighbors (k) to use when picking anchors}
}
\value{
A Seurat object with a new "integrated" Assay. See \code{\link[Seurat:IntegrateData]{Seurat::IntegrateData}} for details.
The object also includes PCA and UMAP reductions and neighbor graphs and seurat clusters defined at res=0.6.
}
\description{
A wrapper function for integration of RNA assays via Seurat using CCA or RPCA; normalizing, regressing out or blacklisting features etc. as specified.
}
\details{
The intention is to streamline analyses where multiple integration methods or parameter combinations need to be compared. Thus, the input Seurat object need only contain raw RNA counts.
}
\examples{
\dontrun{
# log normalize and integrate using canonical correlation analysis dim reduction
s.CCA <- integrate_seurat(s.split, norm_method = "logNorm", reduction = "cca")

# SCTransform and integrate using recripocal PCA dim reduction
s.RPCA <- integrate_seurat(s.split, norm_method = "SCT", dim.reduct = "rpca")

# SCTransform, blacklist TCR genes, and integrate using cca
library(scGate)
TCR_genes <- scGate::genes.blacklist.default$Mm$TCR
s.CCA <- integrate_seurat(s, norm_method = "SCT", dim.reduct = "cca",
                              feature.blacklist = "TCR_genes")
}

}
