% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIntegration.R
\name{integrate_stacas}
\alias{integrate_stacas}
\title{Run integration via STACAS}
\usage{
integrate_stacas(
  s.split,
  nfeatures = 2000,
  npcs = 30,
  norm_method = "logNorm",
  feature.blacklist = NULL,
  regress.out = NULL,
  cell.labels = NULL,
  verbose = FALSE
)
}
\arguments{
\item{s.split}{A split Seurat object containing all samples/captures to integrate}

\item{nfeatures}{Number of features to pass to Seurat::FindVariableFeatures()}

\item{npcs}{Total Number of PCs to compute and store (30 by default)}

\item{norm_method}{Normalization method (either "logNorm" = log transform or "SCT" = SCTransform)}

\item{feature.blacklist}{Vector of genes to blacklist from variable features}

\item{regress.out}{Features to regress out in SCTransform vars.to.regress}

\item{cell.labels}{A metadata column name, storing cell type annotations. These will be taken into account
for semi-supervised alignment (optional). Cells annotated as NA or NULL will not be penalized in semi-supervised
alignment}

\item{verbose}{Print all output}
}
\value{
Integrated Seurat object
}
\description{
A wrapper function for integration of RNA assays via STACAS; normalizing, regressing out or blacklisting features etc. as specified.
}
\details{
The intention is to streamline analyses where multiple integration methods or parameter combinations need to be compared. Thus, the input Seurat object need only contain raw RNA counts.
}
