% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST.R
\name{mastWaldTest}
\alias{mastWaldTest}
\title{Run a Wald test on a MAST fit model}
\usage{
mastWaldTest(zlmFit, hypothesis, baseline, p_adj_method = "fdr")
}
\arguments{
\item{zlmFit}{A MAST model fit produced from \code{MAST::zlm()}}

\item{hypothesis}{A character vector specifying the contrast of interest for the
numerator. Must be compatible with \code{MAST::Hypothesis()}}

\item{baseline}{A MAST hypothesis object specifying the baseline/denominator
for the comparison. Must be compatible with \code{MAST::Hypothesis()}}

\item{p_adj_method}{Method for multiple testing correction of pvalues}
}
\value{
A dataframe of Wald test results
}
\description{
Run a Wald test on a MAST fit model with a specified hypothesis and baseline
to extract P value and LogFoldChange. The function is set up so the
Wald test is run on the \code{hypothesis - baseline} and the logFC is calculated
using the full \code{hypothesis} and \code{baseline} matrices.
}
\examples{
\dontrun{
## These are based on the names of terms in my fit object
contrast1 <- '`(Intercept)`+`stagePost`'
contrast0 <- '`(Intercept)`'
res <- mastWaldTest(hypothesis = contrast1, baseline = contrast0, zlmFit = zlm_fit)
}
}
