% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormFindVarFeatScaleData.R
\name{NormFindVarFeatScaleData}
\alias{NormFindVarFeatScaleData}
\title{Normalize, FindVariableFeatures, ScaleData with optional blacklisting}
\usage{
NormFindVarFeatScaleData(
  s,
  norm_method = "logNorm",
  nfeatures = 2000,
  regress.out = NULL,
  feature.blacklist = NULL,
  verbose = F,
  scale_data = NULL
)
}
\arguments{
\item{s}{Seurat object}

\item{norm_method}{Normalization method (either "logNorm" = log transform or "SCT" = SCTransform)}

\item{nfeatures}{Number of features to pass to Seurat::FindVariableFeatures()}

\item{regress.out}{Features to regress out in SCTransform vars.to.regress}

\item{feature.blacklist}{Vector of genes to blacklist from variable features}

\item{verbose}{Print all messages from wrapped functions}

\item{scale_data}{Logical; whether to run Seurat::ScaleData(). If NULL (default), will scale for log normalization but not for SCTransform.}
}
\value{
Seurat object
}
\description{
Normalize RNA counts with either SCTransform or log normalization; find variable features, with the option of blacklisting sets of genes; and opitonally scaling data.
}
\details{
Each combination of options requires slightly different operations, so this wrapper helps keep code tidy and reproducible.
}
\examples{
\dontrun{
# log normalize, find variable features, scale data
s <- NormFindVarFeatScaleData(s, norm_method = "logNorm")

# SCTransform, find variable features, don't scale data
s <- NormFindVarFeatScaleData(s, norm_method = "SCT")

# log normalize, find variable features blacklisting TCR gene (Mus musculus), scale data
library(scGate)
blacklist <- scGate::genes.blacklist.default$Mm$TCR
s <- NormFindVarFeatScaleData(s, norm_method = "logNorm", feature.blacklist = blacklist)
}
}
