% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntegrationDiagnostics.R
\name{plot_smaller}
\alias{plot_smaller}
\title{Modify a UMAP to fit a ggarranged format better}
\usage{
plot_smaller(p)
}
\arguments{
\item{p}{ggplot object}
}
\value{
ggplot object UMAP with no axes and smaller legend
}
\description{
Removes axes and shrinks legend size of \code{\link[Seurat:DimPlot]{DimPlot()}} or \code{\link[gencoreSC:plot.umap.integrated]{plot.umap.integrated()}}.
}
