% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_arranged.R
\name{plotQC_arranged}
\alias{plotQC_arranged}
\title{Plot both ridges and joint qc plots as a ggarrange object}
\usage{
plotQC_arranged(
  obj.unfilt,
  filtName,
  split_by,
  color_by = "mitoRatio",
  cutoffs,
  title
)
}
\arguments{
\item{obj.unfilt}{Unfiltered Seurat object after running 'addQCmetrics()' and 'addQCfilter()'}

\item{filtName}{metadata column to filter by for the plot}

\item{split_by}{metadata column to split ridge plots by (i.e. usually capture ID or sample ID)}

\item{color_by}{metric with to use for scale_color_gradient (i.e. "mitoRatio" or "riboRatio" or any other metric with limits=c(0,1))}

\item{cutoffs}{named list of cutoff values used in 'addQCmetrics()' for the given filtName}

\item{title}{main title of}
}
\value{
an object of class ggarrange, which is a ggplot or a list of ggplot.
}
\description{
Plot both ridges and joint qc plots as an object of class ggarrange
}
