% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCRidgesJoint.R
\name{plotQC_joint}
\alias{plotQC_joint}
\title{Plot QC nUMI vs nGene}
\usage{
plotQC_joint(
  metadata,
  split_by = "capID",
  cutoffs = NULL,
  color_by = "mitoRato",
  facet_colors = FALSE
)
}
\arguments{
\item{metadata}{Seurat object metadata after running 'addQCmetrics()' and 'addQCfilter()'}

\item{split_by}{Metadata column to split ridge plots by (i.e. usually capture ID or sample ID)}

\item{cutoffs}{Named list of cutoffs}

\item{color_by}{Metric with to use for scale_color_gradient (i.e. "mitoRatio" or "riboRatio" or any other metric with limits=c(0,1))}

\item{facet_colors}{Logical; whether to plot facet colors with ggh4x::facet_wrap2 (experimental)}
}
\value{
An object of class ggarrange, which is a ggplot or a list of ggplot.
}
\description{
Visualize the correlation between genes detected and number of UMIs to determine whether strong presence of cells with low numbers of genes/UMIs
}
