% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_arranged.R
\name{plotQCRidgesJoint}
\alias{plotQCRidgesJoint}
\title{Plot both ridges and joint qc plots as a ggarrange object}
\usage{
plotQCRidgesJoint(
  obj,
  filtName,
  split_by,
  color_by = "mitoRatio",
  cutoffs,
  title,
  mixed_sort = T,
  y.text = FALSE
)
}
\arguments{
\item{obj}{Unfiltered Seurat object after running 'addQCmetrics()' and 'addQCfilter()'}

\item{filtName}{Metadata column to filter by for the plot}

\item{split_by}{Metadata column to split ridge plots by (i.e. usually capture ID or sample ID)}

\item{color_by}{Metric with to use for scale_color_gradient (i.e. "mitoRatio" or "riboRatio" or any other metric with limits=c(0,1))}

\item{cutoffs}{Named list of cutoff values used in 'addQCmetrics()' for the given filtName}

\item{title}{Main title}

\item{mixed_sort}{Whether to use \code{gtools::mixedsort()} to determine level order of \code{split_by}.}
}
\value{
An object of class ggarrange, which is a ggplot or a list of ggplot.
}
\description{
Plot both ridges and joint qc plots as an object of class ggarrange
}
