% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRefMapScoresFacet.R
\name{plotRefMapScoresFacet}
\alias{plotRefMapScoresFacet}
\title{Plot reference mapping calls and scores facetted by cluster}
\usage{
plotRefMapScoresFacet(
  obj,
  label_column,
  label_score_column,
  clusters_column = "seurat_clusters",
  min_proportion = 0,
  ncol = 3,
  n_font_size = 2.5,
  facet_order = NULL
)
}
\arguments{
\item{obj}{A seurat object with columns for cell-type annotations and annotation scores}

\item{label_column}{Column to pull annotations labels}

\item{label_score_column}{Column to pull annotation scores}

\item{clusters_column}{Column to pull cell clusters}

\item{min_proportion}{Minimum proportion of a cluster an annotation must
meet to be plotted}

\item{ncol}{Number of columns for facetting}

\item{n_font_size}{Font size of cluster size labels}

\item{facet_order}{Order of annotation levels to show on X axis}
}
\value{
A ggplot object
}
\description{
Creates violin plots of cell annotation calls and confidence scores, facetted
by query object clusters.
}
\details{
Metadata is pulled from columns in the Seurat
object as specified by the user, so it should be agnostic to SingleR, Azimuth,
or other reference mapping functions that provide labels and scores. You
can optionally specify a minimum proportion threshold to plot, which will
remove data for cell-types appearing in less than X percent of a cluster.
Order of annotation levels appearing on the X axis can be specified with
\code{facet_order}.
}
\examples{
\dontrun{
map_prediction_facetplot(
combined.obj,
label_column = 'predicted.cell_type.lymphTS',
label_score_column = 'predicted.cell_type.score.lymphTS',
min_freq = 0.05
)
}
}
