% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSoupX.R
\name{plotSoupXGeneAdjustments}
\alias{plotSoupXGeneAdjustments}
\title{Plot genes with high degree of contamination correction}
\usage{
plotSoupXGeneAdjustments(
  sc,
  ylim_min = 10,
  xlim_min = 0.05,
  hide_ensembl = FALSE,
  ens_pattern = "^ENS"
)
}
\arguments{
\item{sc}{SoupChannel object created from \code{runSoupX()}}

\item{ylim_min}{Minimum value of 'n'umber of original cells' to plot (y axis)}

\item{xlim_min}{Minimum value of 'portion of cells set to zero' to plot (x axis)}

\item{hide_ensembl}{Whether to include genes with \code{ens_pattern} in the plot}

\item{ens_pattern}{Regex pattern for identifying genes with no gene symbol.
Only used to hide unknown genes from the plot if they aren't of interest}
}
\value{
A plotly object
}
\description{
Uses the results of \code{runSoupX()} to produce a plot of genes with a high
degree of contamination correction through SoupX. This can be defined as genes
with a high portion of cell-level counts set to 0 after running SoupX (\code{plotSoupXGeneAdjustments()}).
Another metric is the portion of reads removed for a given gene (\code{plotSoupXGeneAdjustments2()}).
Ideally genes shown here should be strong cell-type markers,
e.g. hemoglobins or immunoglobulins. For lightweight rendering
and to focus on non-trivial signal, only genes found in a sufficient number
of cells and with a sufficient degree of correction are plotted. You can
also optionally just plot genes with a meaningful signal and exclude genes
with only Ensembl IDs.
}
\examples{
\dontrun{
sc <- runSoupX(tod = tod,
  toc = toc,
  clus = clusters)
soupXGeneAdjustPlot(sc, hide_ensembl = TRUE)
}
}
