% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP_split.R
\name{plotUMAP_split}
\alias{plotUMAP_split}
\title{Plot split umap}
\usage{
plotUMAP_split(
  obj.split,
  reduction = "umap",
  group.by = NULL,
  feature = NULL,
  title.prefix = NULL,
  label.size = 4,
  plot_smaller = T,
  legend = F,
  ...
)
}
\arguments{
\item{obj.split}{Unfiltered Seurat object after running \code{\link[gencoreSC:addQCmetrics]{addQCmetrics()}} and \code{\link[gencoreSC:addQCfilter]{addQCfilter()}}}

\item{reduction}{Reduction to use (default: "umap")}

\item{group.by}{Metadata column to color cells by; passed to \code{\link[Seurat:DimPlot]{DimPlot()}} 'group.by'}

\item{feature}{Metadata column to color cells by; passed to \code{\link[Seurat:FeaturePlot]{FeaturePlot()}} 'features' (but this function doesn't support a vector of features)}

\item{title.prefix}{Title prefix; name of obj (i.e. capture or sample ID) will be appended}

\item{label.size}{Size of DimPlot labels}

\item{plot_smaller}{Whether to plot smaller}

\item{legend}{Whether to plot legend (helpful to remove legend if plotting smaller)}

\item{...}{Arguments passed to \code{\link[gencoreSC:plot_smaller]{plot_smaller()}}}
}
\value{
an object of class ggarrange, which is a ggplot or a list of ggplot.
}
\description{
Plot a split umap from a split SeuratObject
}
