% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntegrationDiagnostics.R
\name{plotUmapIntegrated}
\alias{plotUmapIntegrated}
\title{Plot UMAP of integrated (or merged) data}
\usage{
plotUmapIntegrated(seurat.obj, group.by, title, label.size, legend = TRUE)
}
\arguments{
\item{seurat.obj}{Seurat object}

\item{group.by}{Metadata column to group (color) UMAP by}

\item{title}{Title of plot}

\item{label.size}{Cluster and legend text label size}

\item{legend}{Whether to plot legend}
}
\value{
ggplot object UMAP
}
\description{
Wrapper for DimPlot to plot aspect.ratio=1 plot with some other convenient formatting. Used internally by \code{\link[gencoreSC:plotIntegrationDiagnostics]{PlotIntegrationDiagnostics()}}.
}
