% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST.R
\name{prepare_MAST_obj}
\alias{prepare_MAST_obj}
\title{Converts a Seurat object to SCE prepared for MAST}
\usage{
prepare_MAST_obj(obj, exclusion_patterns = c("^RP[SL]", "^MT"), min.cells = 50)
}
\arguments{
\item{obj}{A processed Seurat object with appropriate metadata for MAST modeling}

\item{exclusion_patterns}{Gene patterns to exclude from object, defaults to
ribosomal and mitochondrial human patterns}

\item{min.cells}{Minimum number of cells a gene must appear in}
}
\value{
An object of class SingleCellExperiment
}
\description{
Converts a processed Seurat object to a SingleCellExperiment object for
use with MAST. Additional filtering is done on genes based on pattern expression
to remove uninteresting genes, and filter for genes expressed in fewer than
\code{min.cells} cells. This function also adds the cellular detection rate as
metadata for modeling, per the MAST paper.
}
\examples{
\dontrun{
obj.sce <- prepare_MAST_obj(pdc_obj, min.cells = 150)
zlm_fit <- zlm(~ cdr + stage + Iga + stage:Iga + (1|individual),
  obj.sce,
  method = 'glmer',
  ebayes = FALSE,
  exprs_values = 'logcounts',
  parallel = TRUE,
  fitArgsD = list(nAGQ=0),
  silent = TRUE)
}
}
