% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBD.R
\name{readBD}
\alias{readBD}
\title{Read BD data}
\usage{
readBD(counts_csv, metadata_csv)
}
\arguments{
\item{counts_csv}{path to csv of UMI count table}

\item{metadata_csv}{path to metadata for UMI count table}
}
\value{
Seurat object
}
\description{
Reads in BD counts and metadata and returns Seurat obj
}
\examples{
s <- readBD(counts_csv = here("data/Combined_Exact-Cells_DBEC_MolsPerCell.csv"), metadata_csv = here("data/Exact-Cells_Sample_Tag_Calls.csv"))
}
