% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCounts10x.R
\name{readCounts}
\alias{readCounts}
\title{Read 10X count data}
\usage{
readCounts(filepath, min.cells = 100, min.features = 10, capID)
}
\arguments{
\item{filepath}{path to UMI count matrix (can be raw or unfiltered)}

\item{min.cells}{minimum number of cells per feature for RNA assay}

\item{min.features}{minimum number of features per cell for RNA assay}

\item{capID}{capture ID label, assigned to 'project' in seurat object metadata}
}
\value{
Seurat object with the number of cells and features removed stored in 'nRemoved' in 'misc' slot
}
\description{
Reads in 10X counts and metadata and returns Seurat obj
}
\note{
To do:
\itemize{
\item add option to not remove multiplets and undetermined droplets
\item currently only accepts RNA or RNA + ADT assay, add ability to handle arbitrary additional assays (e.g. ATAC)
}
}
\examples{
s <- readBD(counts_csv = here("data/Combined_Exact-Cells_DBEC_MolsPerCell.csv"), metadata_csv = here("data/Exact-Cells_Sample_Tag_Calls.csv"))
}
