% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDoubletFinder.R
\name{runDoubletFinder}
\alias{runDoubletFinder}
\title{A wrapper for DoubletFinder functions}
\usage{
runDoubletFinder(
  obj,
  PCs,
  sct,
  cores,
  ground_truth = NULL,
  doublet_rate = 0.075,
  assay = "RNA"
)
}
\arguments{
\item{obj}{A Seurat object for a single capture}

\item{PCs}{Number of PCs to use}

\item{sct}{Is data SCTransformed?}

\item{cores}{Number of cores to use for parallelization}

\item{ground_truth}{Optional: An nCell-length character vector of ground-truth doublet classifications (e.g., "Singlet" or "Doublet") used to gauge performance of logistic regression models trained using pANN vectors during ROC analysis.}

\item{doublet_rate}{Estimated doublet rate, based on sequencing device specifications, library prep, etc.}

\item{assay}{Assay to perform DoubletFinder on, default 'RNA'}
}
\value{
A Seurat object with additional metadata columns, including doublet predictions
and additional data used to rerun DoubletFinder with annotation data if needed
}
\description{
Takes a pre-processes Seurat object for a single capture and performs
DoubletFinder functions. The object is returned with an additional metadata
column of DoubletFinder classification calls. This can optionally be used with
Antibody capture data for hashed studies to provide 'ground truth' calls
and improve the doublet detection algorithm.
}
\examples{
\dontrun{
objs <- lapply(objs, runDoubletFinder,
  PCs = 10,
  sct = FALSE,
  cores = 1)
}
}
