% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIntegration.R
\name{runIntegration}
\alias{runIntegration}
\title{Run integration via Seurat, Harmony, or STACAS}
\usage{
runIntegration(
  s.split,
  integration_method = "merge",
  nfeatures = 2000,
  npcs = 30,
  norm_method = "logNorm",
  feature.blacklist = NULL,
  harmony.group.by.vars,
  regress.out = NULL,
  STACAS.supervision_labels = NULL,
  dim.reduct = "cca",
  harmony_norm_merged = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{s.split}{A split Seurat object containing all samples/captures to integrate}

\item{integration_method}{String specifying integration method; any of c("merge", "seurat", "harmony", "stacas")}

\item{nfeatures}{Number of features to pass to Seurat::FindVariableFeatures()}

\item{npcs}{Total Number of PCs to compute and store (30 by default)}

\item{norm_method}{Normalization method (either "logNorm" = log transform or "SCT" = SCTransform)}

\item{feature.blacklist}{Vector of genes to blacklist from variable features}

\item{harmony.group.by.vars}{Var to remove (passed to group.by.vars in harmony::RunHarmony)}

\item{regress.out}{Features to regress out in SCTransform or ScaleData vars.to.regress}

\item{STACAS.supervision_labels}{Vector of cell labels to pass to supervised integration with STACAS. See \code{magrittr::\link[STACAS:Run.STACAS]{Run.STACAS}} \code{cell.labels} parameter for details.}

\item{dim.reduct}{Reduction to use for Seurat integration ("cca" or "rpca"). see \code{\link[Seurat:FindIntegrationAnchors]{Seurat::FindIntegrationAnchors}} reduction parameter for details.}

\item{harmony_norm_merged}{Whether to normalized merged object or normalize each capture separately (logical).}

\item{verbose}{Print all messages from wrapped functions}
}
\value{
Integrated Seurat object post Seurat::FindClusters() (with Seurat's default resolution = 0.8), ready for plotting UMAPs etc.
The object also includes PCA and UMAP reductions and neighbor graphs and seurat clusters defined at res=0.6.
}
\description{
A wrapper function for integration of RNA assays via \code{\link[Seurat:FindIntegrationAnchors]{Seurat}}, \code{\link[harmony:RunHarmony]{Harmony}}, or \code{\link[STACAS:Run.STACAS]{STACAS}}; normalizing, regressing out or blacklisting features etc. as specified.
}
\details{
The intention is to streamline analyses where multiple integration methods or parameter combinations need to be compared. Thus, the input Seurat object need only contain raw RNA counts.
}
\examples{
\dontrun{
# integrate via Seurat with reciprocal PCA dim reduction method
s.RPCA <- runIntegration(s.split, integration_method = "seurat", norm_method = "logNorm",
                        reduction = "rpca")

# integrate via Harmony, normalizing each sample separately
s.Harmony <- runIntegration(s.split, integration_method = "harmony", norm_method = "logNorm",
                            group.by.vars = "Sample_Name", harmony_norm_merged = FALSE)
}

}
