% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCC.R
\name{scoreCC}
\alias{scoreCC}
\title{Add cell cycle metadata}
\usage{
scoreCC(obj, cc.genes)
}
\arguments{
\item{obj}{Seurat object}

\item{cc.genes}{Vector of gene names for cell cycle genes. If you are starting with Ensembl IDs use gencoreSC::EnsDb2GeneName() first.}
}
\value{
Seurat object with cell cycle metadata (plus normalized and scaled data and PCA reduction)
}
\description{
Calculates phase of cell cycle given a list of cell cycle genes
}
