% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addQCfilter.R
\name{addQCfilter}
\alias{addQCfilter}
\title{Add QC filters}
\usage{
addQCfilter(
  obj,
  split.by = NULL,
  filterName = "input",
  by_threshold = FALSE,
  cutoffs = defaultCutoffs,
  by_outlier = FALSE,
  nmads = 4
)
}
\arguments{
\item{obj}{Seurat object}

\item{split.by}{Metadata column to split by for outlier detection if obj is not already split}

\item{filterName}{Name of new metadata column with logical filter}

\item{by_threshold}{Logical to control whether to filter on thresholds defined in 'cutoffs'}

\item{cutoffs}{Named list of cutoffs}

\item{by_outlier}{Logical to control whether to filter outliers}

\item{nmads}{Median absolute deviation used to define outliers}
}
\value{
Seurat object with new metadata column. If passed a split object, will return split object. Same for a merged/single capture object.
}
\description{
Adds logical metadata column for a given set of arbitrary cutoffs and/or outliers
}
