% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCC.R
\name{checkCC}
\alias{checkCC}
\title{Perform cell cycle check analysis}
\usage{
checkCC(
  obj,
  what2return = c("plot_list", "seurat_obj", "both"),
  cc.genes,
  features2plot = c("Phase", "mitoFr", "riboFr")
)
}
\arguments{
\item{obj}{Seurat object}

\item{what2return}{c("plot_list", "seurat_obj", "both")}

\item{cc.genes}{Vector of gene names for cell cycle genes. If you are starting with Ensembl IDs use gencoreSC::EnsDb2GeneName() first.}

\item{features2plot}{Vector of categorical features to plot PCA for e.g. c("Phase", "mitoRatio", "riboRatio"))}
}
\value{
Seurat object with Phase metadata column, a plot list with PCAs of features specified by features2plot, or both in a list object.
}
\description{
Calculates phase of cell cycle given a list of cell cycle genes
}
\note{
This function draws on code from:
\itemize{
\item https://github.com/hbctraining/scRNA-seq_online/blob/master/lessons/cell_cycle_scoring.md
}

To do:
\itemize{
\item Fix namespace warnings and notes from using methods defined by AnnotationHub
}
}
\examples{
\dontrun{
  # Inferred by orthology searches against Human list published in Tirosh, I, et al.:
  mus_url <- "https://raw.githubusercontent.com/hbc/tinyatlas/master/cell_cycle/Mus_musculus.csv"
  cc_file <- getURL(mus_url)
  cc_cycle_genes <- read.csv(text = cc_file)

  # Only return seurat object with new Phase metadata column
  obj <- checkCC(obj, what2return = "seurat_obj", cc.genes = cc_cycle_genes)

  # Run cell cycle check and only return seurat object
  cc.plots <- checkCC(obj, what2return = "plot_list", cc.genes = cc_cycle_genes)

  # Run cell cycle check and return both
  list[obj, cc.plots] \%<-\% checkCC(obj, what2return = "both", cc.genes = cc_cycle_genes)

}

}
