% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAzimuthReference.R
\name{createAzimuthReference}
\alias{createAzimuthReference}
\title{Create an Azimuth reference and save to file}
\usage{
createAzimuthReference(ref, metadata_column, output_folder, ndims = 50, ...)
}
\arguments{
\item{ref}{A Seurat object with a counts matrix and cell-type annotations}

\item{metadata_column}{Columns of metadata to transfer onto query datasets}

\item{output_folder}{Where to save the Azimuth reference files for future calls
of \code{RunAzimuth}}

\item{ndims}{Number of dimensions to use for PCA and UMAP}

\item{\dots}{Arguments passed to Azimuth::AzimuthReference()}

}
\value{
A Seurat object with AzimuthData stored in the tools slot for use with Azimuth
}
\description{
Create an Azimuth compatible reference from an existing Seurat object
with a counts matrix and cell-type annotations in a metadata column. This
wrapper will process the reference in a standard SCTransform workflow.
The column of metadata with annotations should be specified to the \code{metadata}
parameter. The reference will be saved to the specified output folder to
be referenced in calls to \code{Azimuth::RunAzimuth}.
}
\note{
Code based on scripts from azimuth-references github, such as
https://github.com/satijalab/azimuth-references/blob/master/human_motorcortex/scripts/export.R
}
\examples{
\dontrun{
  ## Load reference data as Seurat object
  ref <- readRDS(here('saved_rds/tabula_sapiens_lymph_subset.Rds'))
  ## No processing needed, plug in
  azimuth_ref <- createAzimuthReference(ref,
                                        'cell_type',
                                        output_folder = here('reference/lymph_node_TS'))

  ## Don't need the reference object for this, just the files
  combined.obj <- RunAzimuth(combined.obj,
                             reference = here('reference/lymph_node_TS'),
                             umap.name = 'refUMAP.lnTS',
                             assay = 'SCT')
}
}
