% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcoreVlnPlot.R
\name{gcoreVlnPlot}
\alias{gcoreVlnPlot}
\title{Violin plot of gene(s) expression grouped by metadata}
\usage{
gcoreVlnPlot(
  obj,
  genes,
  grouping_var,
  groups = NULL,
  subset = NULL,
  subset_var = NULL,
  filter_zeros = TRUE,
  assay = "RNA"
)
}
\arguments{
\item{obj}{Seurat object}

\item{genes}{Genes to plot}

\item{grouping_var}{Column of \code{obj@meta.data} to group data by.}

\item{groups}{Optional. Levels of \code{grouping_var} to include in the plot. Also used
to specify order of levels.}

\item{subset}{Levels of \code{subset_var} to subset data to}

\item{subset_var}{Optional. Column of \code{obj@meta.data} to subset on. Default
is \code{seurat_clusters} so you could \code{subset} on a specific cluster.}

\item{filter_zeros}{Remove 0s from plot: default \code{TRUE}.}

\item{assay}{Assay to pull expression data from, default \code{RNA}}
}
\value{
A ggplot
}
\description{
Similar to \code{Seurat::VlnPlot} but more customizable. Pass a seurat object and
genes to plot. Optionally subset the data by a meta.data variable defined
with \code{subset_var}, selecting level(s) \code{subset}. Useful to plot expression
in one subset of your data without having to make separate objects.
Group the data by meta.data variable \code{grouping_var}, selecting levels \code{groups}.
Optionally filter zeros to focus on cells expressing gene.
}
\details{
Requires \code{ggforce} to be installed
}
\examples{
\dontrun{
## Plots only the 'Intermediate' cells (as labeled in 'coarse_labels' column)
## Groups results by 'Pre' and 'Post', as labeled in the 'stage' column'
  gcoreVlnPlot(myeloid_obj,
    genes = c('ISG15', 'ISG20', 'CD14', 'LDHA', 'IFITM1'),
    subset = 'Intermediate', subset_var = 'coarse_labels',
    grouping_var = 'stage', groups = c('Pre', 'Post'))
}
}
