% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcoreVlnPlot.R
\name{gcoreVlnPlot_facetted}
\alias{gcoreVlnPlot_facetted}
\title{Extension of gcoreVlnPlot intended for facetting data for a single gene}
\usage{
gcoreVlnPlot_facetted(
  obj,
  gene,
  facet_var,
  grouping_var,
  groups = NULL,
  subset = NULL,
  subset_var = NULL,
  filter_zeros = TRUE,
  assay = "RNA"
)
}
\arguments{
\item{gene}{Gene to plot}

\item{facet_var}{Metadata variable to facet plot on}
}
\description{
An extension of gcoreVlnPlot, focusing on a single gene but facetting
it by a metadata variable. The intended use case for this was to facet
by individual/capture to see if a single individual is driving an effect
}
