% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencoreSC-package.R
\docType{package}
\name{gencoreSC-package}
\alias{gencoreSC}
\alias{gencoreSC-package}
\title{gencoreSC: ENPRC Genomics Core Single Cell Utility Functions}
\description{
Convenience functions for single cell analyses.
}
\author{
\strong{Maintainer}: Micah Fletcher \email{mpfletc@emory.edu}

Authors:
\itemize{
  \item Derrik Gratz \email{derrik.gratz@emory.edu} (\href{https://orcid.org/0009-0002-5934-576X}{ORCID}) [contributor]
}

}
\keyword{internal}
