% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIntegration.R
\name{integrate_harmony}
\alias{integrate_harmony}
\title{Run integration via Harmony}
\usage{
integrate_harmony(
  s,
  norm_method = "logNorm",
  nfeatures = 2000,
  npcs = 30,
  group.by.vars = "Sample_Name",
  verbose = FALSE,
  regress.out = NULL,
  feature.blacklist = NULL,
  scale_data = NULL,
  reduction.save = "harmony"
)
}
\arguments{
\item{s}{A Seurat object. If it is a split object, will run normalize samples separately. If merged, will normalize samples together.}

\item{norm_method}{Normalization method (either "logNorm" = log transform or "SCT" = SCTransform)}

\item{nfeatures}{Number of features to pass to Seurat::FindVariableFeatures()}

\item{npcs}{Total Number of PCs to compute and store (30 by default)}

\item{group.by.vars}{Which variable(s) to remove (character vector).}

\item{verbose}{Whether to print progress messages. TRUE to print, FALSE to
suppress.}

\item{regress.out}{Features to regress out in SCTransform vars.to.regress}

\item{feature.blacklist}{Vector of genes to blacklist from variable features}

\item{scale_data}{Logical; whether to run Seurat::ScaleData(). If NULL (default), will scale for log normalization but not for SCTransform.}

\item{reduction.save}{Keyword to save Harmony reduction. Useful if you want
to try Harmony with multiple parameters and save them as e.g.
'harmony_theta0', 'harmony_theta1', 'harmony_theta2'}
}
\value{
Integrated Seurat object
}
\description{
A wrapper function for integration of RNA assays via Harmony; normalizing, regressing out or blacklisting features etc. as specified.
}
\details{
The intention is to streamline analyses where multiple integration methods or parameter combinations need to be compared. Thus, the input Seurat object need only contain raw RNA counts.
}
