% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIntegration.R
\name{integrate_seurat}
\alias{integrate_seurat}
\title{Run integration via Seurat}
\usage{
integrate_seurat(
  s.split,
  method = "cca",
  nfeatures = 2000,
  npcs = 30,
  verbose = FALSE,
  norm_method = "logNorm",
  feature.blacklist = NULL,
  regress.out = NULL,
  k.anchor = 5
)
}
\arguments{
\item{s.split}{A split Seurat object containing all samples/captures to integrate}

\item{method}{Reduction to use for Seurat integration ("cca" or "rpca"). see \code{magrittr::\link[Seurat:FindIntegrationAnchors]{FindIntegrationAnchors}} reduction parameter for details.}

\item{nfeatures}{Number of features to pass to Seurat::FindVariableFeatures()}

\item{npcs}{Total Number of PCs to compute and store (30 by default)}

\item{verbose}{Print progress bars and output}

\item{norm_method}{Normalization method (either "logNorm" = log transform or "SCT" = SCTransform)}

\item{feature.blacklist}{Vector of genes to blacklist from variable features}

\item{regress.out}{Features to regress out in SCTransform vars.to.regress}

\item{k.anchor}{How many neighbors (k) to use when picking anchors}
}
\value{
Integrated Seurat object
}
\description{
A wrapper function for integration of RNA assays via Seurat using CCA or RPCA; normalizing, regressing out or blacklisting features etc. as specified.
}
\details{
The intention is to streamline analyses where multiple integration methods or parameter combinations need to be compared. Thus, the input Seurat object need only contain raw RNA counts.
}
