% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIntegration.R
\name{integrate_stacas}
\alias{integrate_stacas}
\title{Run integration via STACAS}
\usage{
integrate_stacas(
  s.split,
  nfeatures = 2000,
  npcs = 30,
  norm_method = "logNorm",
  feature.blacklist = NULL,
  regress.out = NULL,
  cell.labels = NULL,
  verbose = FALSE
)
}
\arguments{
\item{s.split}{A split Seurat object containing all samples/captures to integrate}

\item{nfeatures}{Number of features to pass to Seurat::FindVariableFeatures()}

\item{npcs}{Total Number of PCs to compute and store (30 by default)}

\item{norm_method}{Normalization method (either "logNorm" = log transform or "SCT" = SCTransform)}

\item{feature.blacklist}{Vector of genes to blacklist from variable features}

\item{regress.out}{Features to regress out in SCTransform vars.to.regress}

\item{cell.labels}{A metadata column name, storing cell type annotations. These will be taken into account
for semi-supervised alignment (optional). Cells annotated as NA or NULL will not be penalized in semi-supervised
alignment}

\item{verbose}{Print all output}
}
\value{
Returns a Seurat object with a new integrated Assay. Also, centered, scaled variable features data are returned in the scale.data slot, and the pca of these batch-corrected scale data in the pca 'reduction' slot. See \code{\link[STACAS:Run.STACAS]{STACAS::Run.STACAS}} for details.
The object also includes PCA and UMAP reductions and neighbor graphs and seurat clusters defined at res=0.6.
}
\description{
A wrapper function for integration of RNA assays via \code{\link[STACAS:Run.STACAS]{STACAS}}; normalizing, regressing out or blacklisting features etc. as specified. The intention is to streamline analyses where multiple integration methods or parameter combinations need to be compared. Thus, the input Seurat object need only contain raw RNA counts.
}
\examples{
\dontrun{
# log normalize, find variable features blacklisting TCR genes (Mus musculus)
library(scGate)
scGate::genes.blacklist.default$Mm \%>\% names()
TCR_genes <- scGate::genes.blacklist.default$Mm$TCR
s.STACAS <- integrate_stacas(s.split, norm_method = "logNorm",
                             feature.blacklist = TCR_genes)

# unsupervised STACAS, blacklisting cc, heatshock, mito, ribo, IFn resp. and TCR genes
blacklist_all <- scGate::genes.blacklist.default$Mm \%>\% unlist() \%>\% unname()
s.STACAS <- integrate_stacas(s.split, norm_method = "logNorm",
                             feature.blacklist = blacklist_all)

# "fully-supervised" STACAS,
# using cell labels from previously running SingleR with the ImmGen dataset
# see https://carmonalab.github.io/STACAS.demo/STACAS.demo.html for details
s.STACAS <- integrate_stacas(s.split, norm_method = "logNorm",
                             feature.blacklist = blacklist_all,
                             cell.labels = "ImmGen.main.pruned.labels")

# "semi-supervised" STACAS, using a subset of cell labels
# see https://carmonalab.github.io/STACAS.demo/STACAS.demo.html for details
Tlike <- c("T cells", "NKT", "NK cells", "ILC", "Tgd")
s.split$Tlike_cells <- ifelse(s.split$ImmGen.main.pruned.labels \%in\% Tlike,
                              s.split$ImmGen.main.pruned.labels,
                              NA)
s.STACAS <- integrate_stacas(s.split, norm_method = "logNorm",
                             feature.blacklist = blacklist_all,
                             cell.labels = "Tlike_cells")
}

}
