% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntegrationDiagnostics.R
\name{plot_smaller}
\alias{plot_smaller}
\title{Modify a UMAP to fit a ggarranged format better}
\usage{
plot_smaller(p, discrete = T, smaller_legend = T)
}
\arguments{
\item{p}{ggplot object}

\item{discrete}{whether to use guide_legend (discrete variables) or guide_colorbar (continuous variables)}

\item{smaller_legend}{whether to plot a smaller legend}
}
\value{
ggplot object UMAP with no axes and smaller legend
}
\description{
Removes axes and shrinks legend size of \code{\link[Seurat:DimPlot]{DimPlot()}} or \code{\link[gencoreSC:plotUmapIntegrated]{plotUmapIntegrated()}}.
}
