% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_joint.R
\name{plotQC_joint}
\alias{plotQC_joint}
\title{Plot QC nUMI vs nGene}
\usage{
plotQC_joint(
  metadata,
  split_by = "capID",
  cutoffs = NULL,
  color_by = "mitoRato"
)
}
\arguments{
\item{metadata}{Seurat object metadata after running 'addQCmetrics()' and 'addQCfilter()'}

\item{split_by}{metadata column to split ridge plots by (i.e. usually capture ID or sample ID)}

\item{cutoffs}{named list of cutoffs}

\item{color_by}{metric with to use for scale_color_gradient (i.e. "mitoRatio" or "riboRatio" or any other metric with limits=c(0,1))}
}
\value{
an object of class ggarrange, which is a ggplot or a list of ggplot.
}
\description{
Plots QC metadata for multiple captures or samples with ggplot2::geom_density_ridges
}
