% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP_split.R
\name{plotUMAP_split}
\alias{plotUMAP_split}
\title{Plot split umap}
\usage{
plotUMAP_split(
  obj.split,
  group.by = "seurat_clusters",
  label.size = 4,
  title.prefix = "seurat_clusters",
  title.size = 8
)
}
\arguments{
\item{obj.split}{Unfiltered Seurat object after running 'addQCmetrics()' and 'addQCfilter()'}

\item{group.by}{metadata column to color cells by; passed to Seurat::DimPlot 'group.by'}

\item{label.size}{size of DimPlot labels}

\item{title.prefix}{title prefix; name of obj (i.e. capture or sample ID) will be appended}

\item{title.size}{size of title; size of axis text will title.size-1}
}
\value{
an object of class ggarrange, which is a ggplot or a list of ggplot.
}
\description{
Plot a split umap from a split SeuratObject
}
