% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCounts10x.R
\name{readCounts10x}
\alias{readCounts10x}
\title{Read 10X count data}
\usage{
readCounts10x(
  capID,
  filepath,
  min.cells = 0,
  min.features = 0,
  strip.suffix = FALSE,
  format = "tsv"
)
}
\arguments{
\item{capID}{capture ID label, assigned to 'project' in seurat object metadata}

\item{filepath}{path to UMI count matrix (can be raw or unfiltered)}

\item{min.cells}{Include features detected in at least this many cells. Will subset the counts matrix as well. To reintroduce excluded features, create a new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many features are detected.}

\item{strip.suffix}{Whether to strip suffix from cellranger cell ID}

\item{format}{Input format. "tsv" (default) uses \code{Seurat::Read10X} to load in for the barcodes.tsv, features.tsv and matrix.mtx in the directory specified by the \code{filepat}h param. "h5" loads the \verb{*feature_bc_matrix.h5} file specifed by the \code{filepath} param.}
}
\value{
An object of class Seurat
}
\description{
Reads in 10X counts and metadata and returns Seurat obj
}
\note{
For \code{format = "tsv"} (default), \code{filepath} must specify the path to the \emph{directory} containing \code{barcodes.tsv}, \code{features.tsv}, and \code{matrix.mtx}; whereas for \code{format = h5}, \code{filepath} must specify the path to the \verb{*feature_bc_matrix.h5} \emph{file}.
}
