% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCounts10x.R
\name{readCounts10x}
\alias{readCounts10x}
\title{Read 10X count data}
\usage{
readCounts10x(
  capID,
  filepath,
  min.cells = 0,
  min.features = 0,
  strip.suffix = TRUE
)
}
\arguments{
\item{capID}{capture ID label, assigned to 'project' in seurat object metadata}

\item{filepath}{path to UMI count matrix (can be raw or unfiltered)}

\item{min.cells}{Include features detected in at least this many cells. Will
subset the counts matrix as well. To reintroduce excluded features, create a
new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many features are
detected.}
}
\value{
An object of class Seurat
}
\description{
Reads in 10X counts and metadata and returns Seurat obj
}
