% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCountsBD.R
\name{readCountsBD}
\alias{readCountsBD}
\title{Read BD data}
\usage{
readCountsBD(counts_csv, metadata_csv)
}
\arguments{
\item{counts_csv}{path to csv of UMI count table}

\item{metadata_csv}{path to metadata for UMI count table}
}
\value{
Seurat object
}
\description{
Reads in BD counts and metadata and returns Seurat obj
}
\note{
To do:
\itemize{
\item add option to not remove multiplets and undetermined droplets
\item allow flexible assays; RNA only, or additional arbitrary assays
\item add working example
}
}
