% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDFDoublets.R
\name{removeDFDoublets}
\alias{removeDFDoublets}
\title{Remove cells classified as doublets by DoubletFinder}
\usage{
removeDFDoublets(obj, classifications_col = "DF_classifications")
}
\arguments{
\item{obj}{An Seurat object with a metadata field for DoubletFinder calls}

\item{classifications_col}{Column of DoubletFinder calls}
}
\value{
An object of class Seurat
}
\description{
Remove cells classified as doublets by DoubletFinder
}
\examples{
\dontrun{
  objs <- lapply(objs, runDoubletFinder, PCs = 10, sct = FALSE, cores = 1)
  objs <- lapply(objs, removeDFDoublets)
}
}
