% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSingleR.R
\name{runSingleR}
\alias{runSingleR}
\title{Wrapper for running singleR}
\usage{
runSingleR(
  obj.seurat,
  ref,
  labels,
  de.method = "classic",
  meta.prefix = "singleR",
  num.threads = BiocParallel::bpnworkers(BiocParallel::MulticoreParam()),
  ...
)
}
\arguments{
\item{obj.seurat}{A Seurat object. Can be a merged, non-integrated set
of all samples, as SingleR can opperate on a per-cell level.}

\item{ref}{A numeric matrix of (usually log-transformed) expression values from a reference dataset,
or a \linkS4class{SummarizedExperiment} object containing such a matrix;
see \code{\link[SingleR]{trainSingleR}} for details.

Alternatively, a list or \linkS4class{List} of SummarizedExperiment objects or numeric matrices containing multiple references.
Row names may be different across entries but only the intersection will be used, see Details.}

\item{labels}{A character vector or factor of known labels for all samples in \code{ref}.

Alternatively, if \code{ref} is a list, \code{labels} should be a list of the same length.
Each element should contain a character vector or factor specifying the label for the corresponding entry of \code{ref}.}

\item{de.method}{Arguments controlling the choice of marker genes used for annotation,
see SingleR::trainSingleR()}

\item{meta.prefix}{Character vector used to name outputs stored in the query object}

\item{num.threads}{Integer scalar specifying the number of threads to use for index building and classification.}

\item{...}{
  Arguments passed on to \code{\link[SingleR:SingleR]{SingleR::SingleR}}
  \describe{
    \item{\code{test}}{A numeric matrix of single-cell expression values where rows are genes and columns are cells.
Alternatively, a \linkS4class{SummarizedExperiment} object containing such a matrix.}
    \item{\code{method}}{Deprecated.}
    \item{\code{clusters}}{A character vector or factor of cluster identities for each cell in \code{test}.
If set, annotation is performed on the aggregated cluster profiles, otherwise it defaults to per-cell annotation.}
    \item{\code{genes,sd.thresh,de.method,de.n,de.args}}{Arguments controlling the choice of marker genes used for annotation, see \code{\link[SingleR]{trainSingleR}}.}
    \item{\code{aggr.ref,aggr.args}}{Arguments controlling the aggregation of the references prior to annotation, see \code{\link[SingleR]{trainSingleR}}.}
    \item{\code{recompute}}{Deprecated and ignored.}
    \item{\code{restrict}}{A character vector of gene names to use for marker selection.
By default, all genes in \code{ref} are used.}
    \item{\code{quantile,fine.tune,tune.thresh,prune}}{Further arguments to pass to \code{\link[SingleR]{classifySingleR}}.}
    \item{\code{assay.type.test}}{An integer scalar or string specifying the assay of \code{test} containing the relevant expression matrix,
if \code{test} is a \linkS4class{SummarizedExperiment} object.}
    \item{\code{assay.type.ref}}{An integer scalar or string specifying the assay of \code{ref} containing the relevant expression matrix,
if \code{ref} is a \linkS4class{SummarizedExperiment} object (or is a list that contains one or more such objects).}
    \item{\code{check.missing}}{Logical scalar indicating whether rows should be checked for missing values (and if found, removed).}
  }}
}
\value{
A list containing the query object with reference annotations added as 3 metadata
columns based on the \code{meta.prefix} and a data frame of the full SingleR results.
}
\description{
Maps labels from a reference dataset onto a query set using SingleR. The
default uses \code{BioCparallel::MulticoreParam} to multi-thread the execution,
but that may need to be disabled using the \code{num.threads} argument if your
machine doesn't support that execution.
}
\examples{
\dontrun{
## Install reference dataset
SeuratData::InstallData('pbmc3k')

singler_out <- runSingleR(obj.seurat = pbmc.merged,
                          ref = pbmc3k.final@assays$RNA@data,
                          labels = pbmc3k.final$seurat_annotations,
                          de.method = 'wilcox',
                          meta.prefix = 'singleR.pbmc3k')
## Spliting the list
pbmc.merged <- singler_out[[1]]
singler_out <- singler_out[[2]]

## Using the `\%<-\%` pipe saves typing and memory
library(zealot)
c(pbmc.merged, singler_out) \%<-\% runSingleR(obj.seurat = pbmc.merged,
                                            ref = pbmc3k.final@assays$RNA@data,
                                            labels = pbmc3k.final$seurat_annotations,
                                            de.method = 'wilcox',
                                            meta.prefix = 'singleR.pbmc3k')

## Using ImmGen data for mouse immune cells
# Prep / load reference
ref.ImmGen <- celldex::ImmGenData()

Get cell ontology names and other info
cl <- ontoProc::getOnto('cellOnto')

}
}
