% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCC.R
\name{scoreCC}
\alias{scoreCC}
\title{Add cell cycle metadata}
\usage{
scoreCC(obj, s.genes, g2m.genes)
}
\arguments{
\item{obj}{Seurat object}

\item{cc.genes}{Named list of vectors of S and G2M gene names. The names must be \code{s.genes} and \code{g2m.genes}. If you are starting with Ensembl IDs see examples for code to convert them to gene names.}
}
\value{
Seurat object with cell cycle metadata (plus normalized and scaled data and PCA reduction)
}
\description{
Calculates phase of cell cycle given a list of cell cycle genes
}
\examples{
\dontrun{
## For human data, it's easy because a list of cc genes, `cc.genes`, are loaded with Seurat automatically
load(Seurat)
seurat_objs <- lapply(seurat_objs, scoreCC, s.genes = cc.genes$s.genes, g2m.genes = cc.genes$g2m.genes) # here seurat_objs is a list of seurat objects
qc_pca_plots <- lapply(seurat_objs, plotPCAs)
ggarrange(plotlist = lapply(qc_pca_plots, function(x) x$Phase))

## For another species, you may need to convert from ensembl ids to gene names, like so:
# Getting cell cycle genes for Mus musculus (see https://github.com/hbctraining/scRNA-seq_online/blob/master/lessons/cell_cycle_scoring.md)
library(RCurl)
library(AnnotationHub)
library(dplyr)

# Inferred by orthology searches against Human list published in Tirosh, I, et al.:
cc_file <- getURL("https://raw.githubusercontent.com/hbc/tinyatlas/master/cell_cycle/Mus_musculus.csv")
cc_cycle_genes <- read.csv(text = cc_file)

# These annotations are Ensemble IDs but we need gene names.
# Connect to AnnotationHub
ah <- AnnotationHub()
# Access the Ensembl database for organism
ahDb <- query(ah, pattern = c("Mus musculus", "EnsDb"), ignore.case = TRUE)
# Acquire the latest annotation files
id <- ahDb \%>\% mcols() \%>\% rownames() \%>\% tail(n=1)
# Download the appropriate Ensembldb databases
edb <- ah[[id]]
# Extract gene-level information from database
annotations <- genes(edb, return.type = "data.frame")
# Select annotations of interest
annotations <- annotations \%>\% dplyr::select(gene_id, gene_name, seq_name, gene_biotype, description)

# Now we can use these annotaitons to get teh corresponding gene names for the Ensembl IDs of the cell cycle genes.
# Get gene names for Ensemble IDs for each gene
cell_cycle_markers <- cc_cycle_genes \%>\% dplyr::left_join(annotations, by =c("geneID" = "gene_id"))

cc_genes <- list()
cc_genes$s.genes <- cell_cycle_markers \%>\% dplyr::filter(phase == "S") \%>\% pull("gene_name")
cc_genes$g2m.genes <- cell_cycle_markers \%>\% dplyr::filter(phase == "G2/M") \%>\% pull("gene_name")

## Now you can run scoreCC and check the PCAs as you would for humans
seurat_objs <- lapply(seurat_objs, scoreCC, s.genes = cc.genes$s.genes, g2m.genes = cc.genes$g2m.genes) # here seurat_objs is a list of seurat objects
qc_pca_plots <- lapply(seurat_objs, plotPCAs)
ggarrange(plotlist = lapply(qc_pca_plots, function(x) x$Phase))
}

}
